Add carousels to Home page and convert project to JavaScript
