+++ agrica-website/src/pages/Services.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Container, Row, Col, Card, Button } from 'react-bootstrap';
 import { Link } from 'react-router-dom';
 
-const Services: React.FC = () => {
+const Services = () => {
   return (
     <div style={{ paddingTop: '100px' }}>
       {/* Hero Section */}
